Update TPM to support LDAA
This commit is very big and I apologize for it.

Here are the key points of this commit:
- All structs from the TPM2B family were update to
support sizes of 32-bit. Relevant files are:
-- CommandDispatchData.h and CommandDispatcher.c: updated command descriptor
sizes from 16-bit to 32-bit
-- Marshal.c and Marshal_fp.h: updated marshaling functions
-- Unmarshal.c and Unmarshal_fp.h: updated marshaling functions
-- TPMB.h and TpmTypes.h: for the updated TPM2B structs

- TcpServer.c and TcpServerPosix.c: increase input and output buffers
from 1MiB to 64MiB.

- AsymmetricCommands.c: Fix bug in Kyber 2Phase key exchange. Alice's
public static key wasn't being passed to the TPM (also relevant
Kyber_2Phase_KEX_fp.h). (I know this should be in a separate commit. I'm
sorry). Update LDAA Sign Commit command user API: the user now has to call the
TPM for each commit it must process due to memory limitations and to keep the
TPMs memory footprint low.

- CryptLDaa.c, CryptLDaa_fp.h and LDaa_SignCommit_fp.h: Update SignCommit
function to an iterative interface instead of a bulk one. Move token link
processing to match the iterative interface. We expect the user to start the
commit processing by genrating a token link and then proceed to calculate the
commit in whatever order they see fit.

- Global.c/.h: Increase actionIoBuffer from 8kiB to 16kiB.

- Implementation.h:
-- Increase number of object handles in the TPM from 3 to 5;
-- Increase context size from 8kiB to 64kiB;
-- Increase command size from 16kiB to 64kiB;
-- Increase response size from 16kiB to 64MiB;
-- Increase NV memory size from 65kiB to 20MiB

+++ src/ContextCommands.c
@@ -73,30 +73,30 @@ TPM2_ContextSave(
 		 )
 {
     TPM_RC          result = TPM_RC_SUCCESS;
-    UINT16          fingerprintSize;    // The size of fingerprint in context
+    UINT32          fingerprintSize;    // The size of fingerprint in context
     // blob.
     UINT64          contextID = 0;      // session context ID
     TPM2B_SYM_KEY   symKey;
     TPM2B_IV        iv;
     TPM2B_DIGEST    integrity;
-    UINT16          integritySize;
+    UINT32          integritySize;
     BYTE            *buffer;
     // This command may cause the orderlyState to be cleared due to
     // the update of state reset data. If the state is orderly and
     // cannot be changed, exit early.
     RETURN_IF_ORDERLY;
-    
+
     // Internal Data Update
-    
+
     // This implementation does not do things in quite the same way as described in
-    // Part 2 of the specification. In Part 2, it indicates that the 
-    // TPMS_CONTEXT_DATA contains two TPM2B values. That is not how this is 
-    // implemented. Rather, the size field of the TPM2B_CONTEXT_DATA is used to 
-    // determine the amount of data in the encrypted data. That part is not 
-    // independently sized. This makes the actual size 2 bytes smaller than 
-    // calculated using Part 2. Since this is opaque to the caller, it is not 
+    // Part 2 of the specification. In Part 2, it indicates that the
+    // TPMS_CONTEXT_DATA contains two TPM2B values. That is not how this is
+    // implemented. Rather, the size field of the TPM2B_CONTEXT_DATA is used to
+    // determine the amount of data in the encrypted data. That part is not
+    // independently sized. This makes the actual size 2 bytes smaller than
+    // calculated using Part 2. Since this is opaque to the caller, it is not
     // necessary to fix. The actual size is returned by TPM2_GetCapabilties().
-    
+
     // Initialize output handle.  At the end of command action, the output
     // handle of an object will be replaced, while the output handle
     // for a session will be the same as input
@@ -114,7 +114,7 @@ TPM2_ContextSave(
 	      {
 		  OBJECT              *object = HandleToObject(in->saveHandle);
 		  ANY_OBJECT_BUFFER   *outObject;
-		  UINT16               objectSize = ObjectIsSequence(object)
+		  UINT32               objectSize = ObjectIsSequence(object)
 						    ? sizeof(HASH_OBJECT) : sizeof(OBJECT);
 		  outObject = (ANY_OBJECT_BUFFER *)(out->context.contextBlob.t.buffer
 						    + integritySize + fingerprintSize);
@@ -233,7 +233,7 @@ TPM2_ContextLoad(
     TPM2B_DIGEST        integrityToCompare;
     TPM2B_DIGEST        integrity;
     BYTE                *buffer;    // defined to save some typing
-    INT32               size;       // defined to save some typing
+    UINT32              size;       // defined to save some typing
     TPM_HT              handleType;
     TPM2B_SYM_KEY       symKey;
     TPM2B_IV            iv;
@@ -247,7 +247,7 @@ TPM2_ContextLoad(
     handleType = HandleGetType(in->context.savedHandle);
     // Get integrity from context blob
     buffer = in->context.contextBlob.t.buffer;
-    size = (INT32)in->context.contextBlob.t.size;
+    size = (UINT32)in->context.contextBlob.t.size;
     result = TPM2B_DIGEST_Unmarshal(&integrity, &buffer, &size);
     if(result != TPM_RC_SUCCESS)
 	return result;
